Land Recent QUIC changes.
Invoke the ack notifier as soon as a packet is revived, instead of
waiting for the final ack, which may never occur.

Merge internal change: 65136251
https://codereview.chromium.org/243933003/


Add an EndToEndTest to verify that the server ignores it when a client
attempts to negotiate an RTT of 0.

Merge internal change: 65131428
https://codereview.chromium.org/242003005/


Don't DCHECK if we get a mismatched QUIC version for a CID in the time
wait list.

Merge internal change: 65116167
https://codereview.chromium.org/243293004/


Fix that removes any receiver-reported revived packets from
pending_retransmissions_ list at the sender, and a couple of other tiny
code cleanups. Also enables LargePostFEC end_to_end test, which now
passes for 10K runs.

Fixes FEC-related bug where sender still tries to retransmit packet that
has been indicated as revived by receiver. Also enables end_to_end FEC
test.

Merge internal change: 65092998
https://codereview.chromium.org/243573006/


Cleanup QuicConfig to combine the client and server hello parsing code
into one code path.

Merge internal change: 65073176
https://codereview.chromium.org/243633006/


Change FEC test in end_to_end_test so that packet loss is not in place
until handshake is finished. (Failure rate of test now at ~0.02%)

Small change to QUIC test.

Merge internal change: 65067487
https://codereview.chromium.org/243903002/


Fix a potential divide by 0 error if the initial round trip time is
specified as 0.

Merge internal change: 65060166
https://codereview.chromium.org/243343003/


Change all QUIC negotiable parameters which were suggestions into fixed
values.

Merge internal change: 65044578
https://codereview.chromium.org/243743004/

R=rch@chromium.org

Review URL: https://codereview.chromium.org/244203003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264986 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2d43c4012859e8e1442a24ea8dd4b4ea2bbb1b8f

+++ net/quic/quic_protocol.h
@@ -60,7 +60,7 @@ const QuicByteCount kMaxPacketSize = 1452;
 
 // Maximum size of the initial congestion window in packets.
 const size_t kDefaultInitialWindow = 10;
-const size_t kMaxInitialWindow = 100;
+const uint32 kMaxInitialWindow = 100;
 
 // Default size of initial flow control window.
 const uint32 kDefaultFlowControlSendWindow = 16 * 1024;  // 16 KB
@@ -70,7 +70,7 @@ const uint32 kDefaultFlowControlSendWindow = 16 * 1024;  // 16 KB
 const size_t kMaxTcpCongestionWindow = 200;
 
 // Don't allow a client to suggest an RTT longer than 15 seconds.
-const size_t kMaxInitialRoundTripTimeUs = 15 * kNumMicrosPerSecond;
+const uint32 kMaxInitialRoundTripTimeUs = 15 * kNumMicrosPerSecond;
 
 // Maximum number of open streams per connection.
 const size_t kDefaultMaxStreamsPerConnection = 100;
